+++ tests/conftest.py
@@ -1,9 +1,7 @@
 from pathlib import Path
 
-import dask.array as da
 import numpy as np
 import pytest
-import xarray as xr
 
 
 @pytest.fixture(scope="session")
@@ -14,13 +12,3 @@ def data_dir() -> Path:
 @pytest.fixture(scope="session")
 def ts_ndvi(data_dir):
     yield (-3000, np.loadtxt(data_dir / "ts-ndvi.txt", dtype="int16"))
-
-
-@pytest.fixture(scope="session")
-def to_da(xx):
-    return xr.DataArray(
-        data=da.from_array(xx.data),
-        dims=xx.dims,
-        coords=xx.coords,
-        attrs=xx.attrs,
-    )
